fix(instructor): move instructor-confine into proxy.ts (Next 16 renamed middleware)
Next.js 16 uses proxy.ts, not middleware.ts — the build errored on both being present. Folded the
instructor route-confinement (redirect off /dashboard, /admin/*, student routes → /instructor/dashboard;
authed instructors on auth pages → /instructor/dashboard) into the existing proxy.ts and removed the
stray middleware.ts.
